rename MeshGeoTools namespace to MeshGeoToolsLib to make it consistent with a directory name
